Dedupe `getBoroName` and `getBoroNameMemoized` definitions, see https://github.com/josephfrazier/reported-web/pull/410#pullrequestreview-1327493910
